+++ custom_components/solarcharger/state_machine/state_tidyup.py
@@ -7,6 +7,7 @@ from ..chargers.chargeable import Chargeable
 from ..chargers.charger import Charger
 from ..const import (
     SENSOR_CHARGER_ALLOCATED_POWER,
+    SENSOR_MEDIAN_ALLOCATED_POWER,
     SENSOR_MOVING_AVERAGE_ALLOCATED_POWER,
     RunState,
 )
@@ -36,6 +37,7 @@ class StateTidyUp(SolarChargeState):
         # Unsubscribe allocated power
         self.solarcharge.give_up_real_power_allocation()
         self.solarcharge.tracker.untrack_allocated_power_update()
+        self.solarcharge.entities.sensors[SENSOR_MEDIAN_ALLOCATED_POWER].set_state(0)
         self.solarcharge.entities.sensors[
             SENSOR_MOVING_AVERAGE_ALLOCATED_POWER
         ].set_state(0)